+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Fixed `plugins uninstall` command failing when there is a space in the install path.
+
 ## `4.18.4`
 
 - BugFix: Added the ability to read option values from aliases. Enhanced backward compatibility with V1 profiles.